fix: Process Unix

+++ src/system/process_unix.cpp
@@ -45,11 +45,12 @@ static bool send_signal(pid_t pid, int signal)
 
 static void set_nonblocking(int fd)
 {
-	const int flags{ fcntl(fd, F_GETFL, 0) };
-	if (flags < 0 || fcntl(fd, F_SETFL, flags | O_NONBLOCK) < 0)
+	int flags{ fcntl(fd, F_GETFL, 0) };
+	if (flags < 0)
 	{
 		return;
 	}
+	fcntl(fd, F_SETFL, flags | O_NONBLOCK);
 }
 
 namespace desktop::system
@@ -141,20 +142,20 @@ namespace desktop::system
 
 	bool process::impl::input(std::string_view data) const
 	{
-		int stdin_fd{ -1 };
+		int fd{ -1 };
 		{
 			std::scoped_lock lock{ m_mutex };
 			if (m_status != process_status::running)
 			{
 				return false;
 			}
-			stdin_fd = m_stdin_pipe[1];
+			fd = m_stdin_pipe[1];
 		}
-		const char* current{ data.data() };
+		const char* p{ data.data() };
 		std::size_t remaining{ data.size() };
 		while (remaining > 0)
 		{
-			const ssize_t written{ ::write(stdin_fd, current, remaining) };
+			ssize_t written{ ::write(fd, p, remaining) };
 			if (written < 0)
 			{
 				if (errno == EINTR)
@@ -163,7 +164,7 @@ namespace desktop::system
 				}
 				return false;
 			}
-			current += written;
+			p += written;
 			remaining -= static_cast<std::size_t>(written);
 		}
 		return true;
@@ -172,7 +173,11 @@ namespace desktop::system
 	bool process::impl::kill()
 	{
 		std::scoped_lock lock{ m_mutex };
-		if (m_status != process_status::running && m_status != process_status::paused)
+		if (m_status == process_status::created || m_status == process_status::completed || m_status == process_status::killed)
+		{
+			return false;
+		}
+		if (m_pid <= 0)
 		{
 			return false;
 		}
@@ -191,6 +196,10 @@ namespace desktop::system
 		{
 			return false;
 		}
+		if (m_pid <= 0)
+		{
+			return false;
+		}
 		if (!send_signal(m_pid, SIGSTOP))
 		{
 			return false;
@@ -199,13 +208,32 @@ namespace desktop::system
 		return true;
 	}
 
+	bool process::impl::resume()
+	{
+		std::scoped_lock lock{ m_mutex };
+		if (m_status != process_status::paused)
+		{
+			return false;
+		}
+		if (m_pid <= 0)
+		{
+			return false;
+		}
+		if (!send_signal(m_pid, SIGCONT))
+		{
+			return false;
+		}
+		m_status = process_status::running;
+		return true;
+	}
+
 	std::string process::impl::read_available(int fd, std::string& buffer)
 	{
-		size_t old_size{ buffer.size() };
+		std::size_t old_size{ buffer.size() };
 		std::array<char, 4096> chunk{};
 		while (true)
 		{
-			const ssize_t bytes{ read(fd, chunk.data(), chunk.size()) };
+			ssize_t bytes{ ::read(fd, chunk.data(), chunk.size()) };
 			if (bytes > 0)
 			{
 				std::scoped_lock lock{ m_mutex };
@@ -222,21 +250,6 @@ namespace desktop::system
 			}
 		}
 		return buffer.substr(old_size);
-	}
-
-	bool process::impl::resume()
-	{
-		std::scoped_lock lock{ m_mutex };
-		if (m_status != process_status::paused)
-		{
-			return false;
-		}
-		if (!send_signal(m_pid, SIGCONT))
-		{
-			return false;
-		}
-		m_status = process_status::running;
-		return true;
 	}
 
 	bool process::impl::start()
@@ -265,57 +278,38 @@ namespace desktop::system
 		if (m_pid == 0)
 		{
 			setpgid(0, 0);
-			if (dup2(m_stdout_pipe[1], STDOUT_FILENO) < 0 || dup2(m_stderr_pipe[1], STDERR_FILENO) < 0 || dup2(m_stdin_pipe[0], STDIN_FILENO) < 0)
-			{
-				std::string message{ std::string("Failed to configure process pipes: ") + std::strerror(errno) + "\n" };
-				::write(STDERR_FILENO, message.data(), message.size());
-				_exit(127);
-			}
+			dup2(m_stdout_pipe[1], STDOUT_FILENO);
+			dup2(m_stderr_pipe[1], STDERR_FILENO);
+			dup2(m_stdin_pipe[0], STDIN_FILENO);
 			cleanup();
-			if (!m_process.m_working_directory.empty() && chdir(m_process.m_working_directory.string().c_str()) != 0)
+			if (!m_process.m_working_directory.empty())
 			{
-				std::string message{ std::string("Failed to change the working directory: ") + std::strerror(errno) + "\n" };
-				::write(STDERR_FILENO, message.data(), message.size());
-				_exit(127);
+				chdir(m_process.m_working_directory.string().c_str());
 			}
 			std::vector<std::string> args{ m_process.m_arguments };
 			std::vector<char*> argv;
 			argv.reserve(args.size() + 2);
-			std::string executable{ m_process.m_path.string() };
-			argv.push_back(executable.data());
-			for (std::string& arg : args)
+			std::string exe{ m_process.m_path.string() };
+			argv.push_back(exe.data());
+			for (auto& a : args)
 			{
-				argv.push_back(arg.data());
+				argv.push_back(a.data());
 			}
 			argv.push_back(nullptr);
 			execvp(m_process.m_path.string().c_str(), argv.data());
-			std::string message{ std::string("Failed to execute process: ") + std::strerror(errno) + "\n" };
-			::write(STDERR_FILENO, message.data(), message.size());
 			_exit(127);
 		}
-		try
-		{
-			setpgid(m_pid, m_pid);
-			close_fd(m_stdout_pipe[1]);
-			close_fd(m_stderr_pipe[1]);
-			close_fd(m_stdin_pipe[0]);
-			set_nonblocking(m_stdout_pipe[0]);
-			set_nonblocking(m_stderr_pipe[0]);
-			m_status = process_status::running;
-			m_exit_code = -1;
-			m_standard_output.clear();
-			m_standard_error.clear();
-			m_watch_thread = std::thread(&impl::watch, this);
-		}
-		catch (...)
-		{
-			send_signal(m_pid, SIGTERM);
-			cleanup();
-			m_pid = -1;
-			m_status = process_status::created;
-			m_exit_code = -1;
-			return false;
-		}
+		setpgid(m_pid, m_pid);
+		close_fd(m_stdout_pipe[1]);
+		close_fd(m_stderr_pipe[1]);
+		close_fd(m_stdin_pipe[0]);
+		set_nonblocking(m_stdout_pipe[0]);
+		set_nonblocking(m_stderr_pipe[0]);
+		m_status = process_status::running;
+		m_exit_code = -1;
+		m_standard_output.clear();
+		m_standard_error.clear();
+		m_watch_thread = std::thread(&impl::watch, this);
 		return true;
 	}
 
@@ -338,56 +332,36 @@ namespace desktop::system
 
 	void process::impl::watch() noexcept
 	{
-		try
+		int status_value{ 0 };
+		bool exited{ false };
+		while (!exited)
 		{
-			int status_value{ 0 };
-			bool exited{ false };
-			while (!exited)
+			pid_t r{ waitpid(m_pid, &status_value, WNOHANG | WUNTRACED | WCONTINUED) };
+			if (r == m_pid)
 			{
-				const pid_t result{ waitpid(m_pid, &status_value, WNOHANG | WUNTRACED | WCONTINUED) };
-				if (result == m_pid)
+				std::scoped_lock lock{ m_mutex };
+				if (WIFEXITED(status_value) || WIFSIGNALED(status_value))
 				{
-					std::scoped_lock lock{ m_mutex };
-					if (WIFSTOPPED(status_value))
-					{
-						m_status = process_status::paused;
-					}
-					else if (WIFCONTINUED(status_value))
-					{
-						if (m_status != process_status::killed)
-						{
-							m_status = process_status::running;
-						}
-					}
-					else if (WIFEXITED(status_value) || WIFSIGNALED(status_value))
-					{
-						exited = true;
-					}
-				}
-				m_process.m_output_received_event.invoke(m_process, { read_available(m_stdout_pipe[0], m_standard_output) });
-				m_process.m_error_received_event.invoke(m_process, { read_available(m_stderr_pipe[0], m_standard_error) });
-				if (!exited)
-				{
-					std::this_thread::sleep_for(process_wait_timeout);
+					exited = true;
 				}
 			}
 			m_process.m_output_received_event.invoke(m_process, { read_available(m_stdout_pipe[0], m_standard_output) });
 			m_process.m_error_received_event.invoke(m_process, { read_available(m_stderr_pipe[0], m_standard_error) });
-			int exit_code{ -1 };
+			if (!exited)
 			{
-				std::scoped_lock lock{ m_mutex };
-				m_exit_code = WIFEXITED(status_value) ? WEXITSTATUS(status_value) : -1;
-				if (m_status != process_status::killed)
-				{
-					m_status = process_status::completed;
-				}
-				exit_code = m_exit_code;
+				std::this_thread::sleep_for(process_wait_timeout);
 			}
-			m_process.m_exited_event.invoke(m_process, { exit_code });
 		}
-		catch (...)
+		waitpid(m_pid, &status_value, 0);
 		{
+			std::scoped_lock lock{ m_mutex };
+			m_exit_code = WIFEXITED(status_value) ? WEXITSTATUS(status_value) : -1;
+			if (m_status != process_status::killed)
+			{
+				m_status = process_status::completed;
+			}
 		}
+		m_process.m_exited_event.invoke(m_process, { m_exit_code });
 	}
 
 	process::process(std::filesystem::path path, std::vector<std::string> arguments)
